feat: add height and age input steps and gender selection menu

+++ Models/GoalSetup.cs
@@ -3,7 +3,7 @@ namespace NutriBeastBot.Models;
 public class GoalSetup
 {
     public double Weight { get; set; }
-    public int Height { get; set; }
+    public double Height { get; set; }
     public int Age { get; set; }
     public string Gender { get; set; } = "";
     public string Goal { get; set; } = "";
